refactor(grpc): rename and split service modules
