+++ rust/src/server/src/grpc/health.rs
@@ -8,14 +8,14 @@ use tonic_health::ServingStatus;
 use tonic_health::server::HealthReporter;
 use tracing::{info, warn};
 
-use super::{ControlGrpcService, GenerateGrpcService};
+use super::{ControlGrpcService, InferenceGrpcService};
 
 pub(crate) async fn monitor_health(
     mut health_reporter: HealthReporter,
     mut engine_health: watch::Receiver<bool>,
     shutdown: CancellationToken,
 ) {
-    let generate_service = GenerateGrpcService::NAME;
+    let inference_service = InferenceGrpcService::NAME;
     let control_service = ControlGrpcService::NAME;
     let status = ServingStatus::NotServing;
     let health_event_first = tokio::select! {
@@ -45,7 +45,7 @@ pub(crate) async fn monitor_health(
         }
     };
 
-    health_reporter.set_not_serving::<GenerateGrpcService>().await;
+    health_reporter.set_not_serving::<InferenceGrpcService>().await;
     health_reporter.set_not_serving::<ControlGrpcService>().await;
     // Both gRPC services use the same engine client, so overall server health
     // mirrors their shared engine health.
@@ -59,7 +59,7 @@ pub(crate) async fn monitor_health(
         );
     }
 
-    health_reporter.clear_service_status(generate_service).await;
+    health_reporter.clear_service_status(inference_service).await;
     health_reporter.clear_service_status(control_service).await;
     health_reporter.clear_service_status("").await;
 }
